requirements.txt
# Streamlit Cloud Requirements
streamlit>=1.28.0
pandas>=1.5.0
psycopg2-binary>=2.9.0
python-dotenv>=0.19.0
plotly>=5.17.0
altair>=5.0.0
requests>=2.31.0
urllib3>=2.0.0
openai>=1.0.0
